refactor: extract types to separate file
